requirements.txt
# Backend Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Search and Text Processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# File Handling
python-magic==0.4.27
chardet==5.2.0

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
click==8.1.7
colorama==0.4.6
apscheduler==3.10.4

# Development/Testing
pytest==7.4.3
pytest-flask==1.3.0

# Monitoring and Logging
python-json-logger==2.0.7

# Performance
redis==5.0.1
gunicorn==21.2.0
